Add basic template engine
This adds basic template parsing and processing using what looks like a Visitor pattern.

I'm not 100% sure, as I'm not classically trained. But this looks about right?

I think there's some stuff I'm doing that doesn't make a lot of sense, but that
might be the abstraction at work.

It works though, which is neat.

All this does it take the input string, and then outputs it again. Everything is joined
together to output a string at the end.

Next up is adding placeholder values that can be replaced with provided data, which will make
it more useful. Probably!