Add placeholder processing
The template engine is now able to process placeholder values that are defined in
curly braces in the template, and replace them with values contained within a map
of data that is passed in when processing.

There's no real entrypoint yet for how this is going to work, but I expect
we're likely to want to read this from a yaml file of some sort.